# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ZTF Foundation API Documentation",
        "description": "API Documentation for ZTF Foundation User Management",
        "contact": {
            "name": "ZTF Foundation Team",
            "email": "admin@ztffoundation.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "ZTF Foundation API Server"
        }
    ],
    "paths": {
        "/api/users": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get list of users",
                "operationId": "862e982d8596a916e4f030cbce92bc53",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer"
                                            },
                                            "name": {
                                                "type": "string"
                                            },
                                            "email": {
                                                "type": "string"
                                            },
                                            "matricule": {
                                                "type": "string"
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "datetime"
                                            },
                                            "updated_at": {
                                                "type": "string",
                                                "format": "datetime"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/users/{id}": {
            "put": {
                "tags": [
                    "Users"
                ],
                "summary": "Update user information",
                "operationId": "7b1d5722e544300ce266edb80c6cac66",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "matricule",
                                    "email"
                                ],
                                "properties": {
                                    "matricule": {
                                        "type": "string",
                                        "example": "CM-HQ-IT-001"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User updated successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Users"
                ],
                "summary": "Delete a user",
                "operationId": "fe686921a22e8518e2b486057ab1863b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User deleted successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "User not found"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/register": {
            "get": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Affiche le formulaire d'inscription final",
                "description": "Retourne la vue du formulaire d'inscription",
                "operationId": "250c656d98c15d552eaf8f9310b5fda3",
                "responses": {
                    "200": {
                        "description": "Vue du formulaire d'inscription"
                    }
                }
            }
        },
        "/users": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Liste tous les utilisateurs",
                "description": "Retourne la liste de tous les utilisateurs avec leurs services et départements",
                "operationId": "7e2f5799553ec93c9c43e2a58490d447",
                "responses": {
                    "200": {
                        "description": "Liste des utilisateurs récupérée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer"
                                            },
                                            "name": {
                                                "type": "string"
                                            },
                                            "email": {
                                                "type": "string"
                                            },
                                            "matricule": {
                                                "type": "string"
                                            },
                                            "services": {
                                                "type": "array",
                                                "items": {
                                                    "type": "object"
                                                }
                                            },
                                            "departments": {
                                                "type": "array",
                                                "items": {
                                                    "type": "object"
                                                }
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/users/{user}": {
            "put": {
                "tags": [
                    "Users"
                ],
                "summary": "Met à jour les informations d'un utilisateur",
                "description": "Met à jour le matricule, l'email et le mot de passe d'un utilisateur",
                "operationId": "caaec06bf33a577d630c028222869ee6",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "ID de l'utilisateur",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "matricule",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "matricule": {
                                        "type": "string",
                                        "example": "CM-HQ-IT-001"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Utilisateur mis à jour avec succès"
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Users"
                ],
                "summary": "Supprime un utilisateur",
                "description": "Supprime définitivement un utilisateur de la base de données",
                "operationId": "a4e18f8517afcc062825920ffb6cd762",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "ID de l'utilisateur à supprimer",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Utilisateur supprimé avec succès"
                    },
                    "404": {
                        "description": "Utilisateur non trouvé"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Users",
            "description": "API Endpoints for User Management"
        },
        {
            "name": "Authentication",
            "description": "Authentication"
        }
    ]
}